requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
PyMySQL==1.1.1
pandas==2.3.3
numpy==2.3.4
scikit-learn==1.7.2
pypinyin==0.53.0
cryptography==45.0.6
